feat : Post 업데이트 및 삭제시 UserId 확인하도록 수정

+++ raonz/src/main/java/hgu/se/raonz/post/application/service/PostService.java
@@ -23,10 +23,10 @@ public class PostService {
     private final UserRepository userRepository;
 
     @Transactional
-    public Post addPost(PostRequest postRequest, int type, String user_id) {
-        List<User> userList = userRepository.findUserListByUserId(user_id);
+    public Post addPost(PostRequest postRequest, int type, String userId) {
+        List<User> userList = userRepository.findUserListByUserId(userId);
         if (userList.size() != 1) {
-            System.out.println(user_id);
+            System.out.println(userId);
             System.out.println("userList size = " + userList.size() + "Duplicate email");
 //            return null;
         }
@@ -38,10 +38,15 @@ public class PostService {
     }
 
     @Transactional
-    public Long deletePost(Long postId) {
+    public Long deletePost(Long postId, String userId) {
         Post post = postRepository.findById(postId).orElse(null);
+        List<User> userList = userRepository.findUserListByUserId(userId);
 
         if(post != null) {
+            if (!post.getUser().getUserId().equals(userId)) { // 나중에 팀 클래스 생기면 수정 필요
+                System.out.println(userId);
+                return null;
+            }
             Long returnId = post.getPostId();
             postRepository.delete(post);
 
@@ -52,10 +57,14 @@ public class PostService {
     }
 
     @Transactional
-    public Long updatePost(Long postId, PostUpdateRequest postUpdateRequest) {
+    public Long updatePost(Long postId, PostUpdateRequest postUpdateRequest, String userId) {
         Post post = postRepository.findById(postId).orElse(null);
 
         if(post != null) {
+            if (!post.getUser().getUserId().equals(userId)) { // 나중에 팀 클래스 생기면 수정 필요
+                System.out.println(userId);
+                return null;
+            }
             post.setTitle(postUpdateRequest.getTitle());
             post.setContent(postUpdateRequest.getContent());
             postRepository.save(post);
